Set up application with an application factory
This application factory is a function that is used to create an
instance of the Flask class needed to create a new Flask application.
Any configuration, registration, and other setup the application needs
will happen inside this function, and then the application will be
returned.

This application set up was taken directly from the Flask tutorial
found at: https://flask.palletsprojects.com/en/3.0.x/tutorial/factory/