Fix bugs in login view
Previously, the database query to get the user information was not
working correctly due to a lack of spaces in the query. Additionally,
the login form wasn't being displayed previously due to a typo in the
path to "login.html" as an argument to render_template(). Both bugs have
been fixed and now work as intended.