Import into database anime data obtained from the AnimeNewsNetwork API
Also updated the schema.sql file to include a new column for the
anime_shows table in the database to keep track of the precision for
each anime's data. An anime can have different entries in the
AnimeNewsNetwork database based on its number of seasons. The precision
helps to indicate what season a certain entry is for an anime show.