Create search view and its html templates
The user will be able to submit a form in search.html to search for
an anime by its title in the database. The result will then be
displayed in searched.html. A link to this new search view was also
included in the base html template, as well as a link to the index view.