Create details view and its html template
After searching for an anime by its title, the results shown in
searched.html will each redirect to the details view, which will
display a list of that anime's different releases in details.html. The
full title of each release will be shown in the list and each release
will have a link that links to that release's page in the
AnimeNewsNetwork website that displays further details for that release.