Create edit view
Users can now edit information about the anime in their collection,
such as how much an anime cost them and when they bought it. An edit
page will be displayed whenever a user first adds an anime release
to their collection, and it can be accessed again through the index
page where a user's collection is displayed.